NEW Refactor diff API and add managed limits.
Deprecate the getDiffedChanges method to get rid of confusing
fullHistory parameter. Add a getDiff method to fetch just a single diff
and a more versatile getDiffList method which allows specifying a limit.

We now also cap the amount of items coming from the diffing process
to a default value that we hope will result in generation times no
greater than a few seconds. This can be reconfigured by developers.